#342 Organization service release, layout update

+++ MiniSpace.Services.MediaFiles/src/MiniSpace.Services.MediaFiles.Application/Exceptions/InvalidFileSizeException.cs
@@ -3,10 +3,10 @@
     public class InvalidFileSizeException : AppException
     {
         public override string Code { get; } = "invalid_file_size";
-        public int FileSize { get; }
-        public int MaxFileSize { get; }
+        public long FileSize { get; }
+        public long MaxFileSize { get; }
 
-        public InvalidFileSizeException(int fileSize, int maxFileSize) 
+        public InvalidFileSizeException(long fileSize, long maxFileSize) 
             : base($"Invalid file size: {fileSize}. Maximum valid file size: {maxFileSize}.")
         {
             FileSize = fileSize;
